+++ Changelog.rst
@@ -1,6 +1,22 @@
 Change Log
 ============
 
+1.7.0
+++++++
+
+Changes
+--------
+
+* Better tunneling implementation for native clients that supports multiple tunnels over single SSH connection for connecting multiple hosts through single proxy.
+* Added ``greenlet_timeout`` setting to native client ``run_command`` to pass on to getting greenlet result to allow for greenlets to timeout.
+
+
+Fixes
+------
+
+* Native client tunneling would not work correctly - #123.
+* ``timeout`` setting was not applied to native client sockets.
+
 1.6.3
 ++++++